# api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/flight/track": {
      "get": {
        "description": "Get the flight path with all position data points for a specific flight",
        "operationId": "FlightController_getFlightTrack_v1",
        "parameters": [
          {
            "name": "iata",
            "required": true,
            "in": "query",
            "description": "IATA flight number",
            "schema": {
              "example": "UA123",
              "type": "string"
            }
          },
          {
            "name": "icao",
            "required": true,
            "in": "query",
            "description": "ICAO flight number",
            "schema": {
              "example": "UAL123",
              "type": "string"
            }
          },
          {
            "name": "date",
            "required": false,
            "in": "query",
            "description": "Flight date (ISO format). Defaults to current date if not provided",
            "schema": {
              "example": "2024-03-20",
              "type": "string"
            }
          },
          {
            "name": "timezone",
            "required": true,
            "in": "query",
            "description": "The timezone from which the request is being made.",
            "schema": {
              "example": "IST",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Flight track data retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FlightTrackResponseEntity"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request parameters or flight not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericErrorEntity"
                }
              }
            }
          },
          "401": {
            "description": "User is not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericErrorEntity"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericErrorEntity"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get flight track/path",
        "tags": [
          "Flight"
        ]
      }
    },
    "/v1/flight": {
      "get": {
        "description": "Get detailed information about a specific flight including real-time tracking data",
        "operationId": "FlightController_getFlight_v1",
        "parameters": [
          {
            "name": "iata",
            "required": true,
            "in": "query",
            "description": "IATA flight number",
            "schema": {
              "example": "UA123",
              "type": "string"
            }
          },
          {
            "name": "icao",
            "required": true,
            "in": "query",
            "description": "ICAO flight number",
            "schema": {
              "example": "UAL123",
              "type": "string"
            }
          },
          {
            "name": "date",
            "required": false,
            "in": "query",
            "description": "Flight date (ISO format). Defaults to current date if not provided",
            "schema": {
              "example": "2024-03-20",
              "type": "string"
            }
          },
          {
            "name": "timezone",
            "required": true,
            "in": "query",
            "description": "The timezone from which the request is being made.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "forceUpdate",
            "required": false,
            "in": "query",
            "description": "Force update the flight data(used to refresh the flight data)",
            "schema": {
              "default": false,
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Flight details retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FlightResponseEntity"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericErrorEntity"
                }
              }
            }
          },
          "401": {
            "description": "User is not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericErrorEntity"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericErrorEntity"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get flight details",
        "tags": [
          "Flight"
        ]
      }
    },
...
  },
  "info": {
    "title": "Plane Pal",
    "description": "Plane Pal API Documentation",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "GenericErrorEntity": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        },
        "required": [
          "message"
        ]
      },
...
      "AircraftEntity": {
        "type": "object",
        "properties": {
          "modeS": {
            "type": "string"
          },
          "registration": {
            "type": "string"
          },
          "model": {
            "type": "string"
          },
          "image": {
            "type": "string"
          },
          "aircraft_id": {
            "type": "string"
          },
          "age": {
            "type": "string"
          },
          "firstFlightDate": {
            "type": "string"
          },
          "deliveryDate": {
            "type": "string"
          },
          "isFreighter": {
            "type": "boolean"
          },
          "attribution": {
            "type": "string"
          },
          "payload": {
            "type": "object"
          }
        }
      },
      "PartialAirlineEntity": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "iata": {
            "type": "string"
          },
          "icao": {
            "type": "string"
          },
          "image": {
            "type": "string"
          }
        },
        "required": [
          "name"
        ]
      },
      "LocationEntity": {
        "type": "object",
        "properties": {
          "lat": {
            "type": "number"
          },
          "lon": {
            "type": "number"
          }
        },
        "required": [
          "lat",
          "lon"
        ]
      },
      "RouteAirportEntity": {
        "type": "object",
        "properties": {
          "iata": {
            "type": "string"
          },
          "icao": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "timeZone": {
            "type": "string"
          },
          "shortName": {
            "type": "string"
          },
          "countryCode": {
            "type": "string"
          },
          "municipalityName": {
            "type": "string"
          },
          "location": {
            "$ref": "#/components/schemas/LocationEntity"
          }
        },
        "required": [
          "iata",
          "icao",
          "name",
          "timeZone",
          "countryCode",
          "municipalityName",
          "location"
        ]
      },
      "TimeEntity": {
        "type": "object",
        "properties": {
          "utc": {
            "type": "string"
          },
          "local": {
            "type": "string"
          }
        },
        "required": [
          "utc",
          "local"
        ]
      },
      "RouteInfoEntity": {
        "type": "object",
        "properties": {
          "airport": {
            "$ref": "#/components/schemas/RouteAirportEntity"
          },
          "terminal": {
            "type": "string"
          },
          "gate": {
            "type": "string"
          },
          "scheduledTime": {
            "$ref": "#/components/schemas/TimeEntity"
          },
          "revisedTime": {
            "$ref": "#/components/schemas/TimeEntity"
          },
          "predictedTime": {
            "$ref": "#/components/schemas/TimeEntity"
          },
          "baggageBelt": {
            "type": "string"
          }
        },
        "required": [
          "airport",
          "scheduledTime"
        ]
      },
      "StringifiedGreatCircleDistanceEntity": {
        "type": "object",
        "properties": {
          "meter": {
            "type": "string"
          },
          "km": {
            "type": "string"
          },
          "mile": {
            "type": "string"
          },
          "nm": {
            "type": "string"
          },
          "feet": {
            "type": "string"
          }
        }
      },
      "FlightAwareDataEntity": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "ident": {
            "type": "string",
            "description": "Flight identifier"
          },
          "identIcao": {
            "type": "string"
          },
          "identIata": {
            "type": "string"
          },
          "actualRunwayOff": {
            "type": "string"
          },
          "actualRunwayOn": {
            "type": "string"
          },
          "operator": {
            "type": "string"
          },
          "operatorIcao": {
            "type": "string"
          },
          "operatorIata": {
            "type": "string"
          },
          "flightNumber": {
            "type": "string"
          },
          "registration": {
            "type": "string"
          },
          "atcIdent": {
            "type": "string"
          },
          "inboundFaFlightId": {
            "type": "string"
          },
          "codesharesIata": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "departureDelay": {
            "type": "number",
            "description": "Departure delay in seconds (negative = early)"
          },
          "arrivalDelay": {
            "type": "number",
            "description": "Arrival delay in seconds (negative = early)"
          },
          "filedEte": {
            "type": "number",
            "description": "Estimated time en route in seconds"
          },
          "progressPercent": {
            "type": "number",
            "description": "Percent completion (0-100)"
          },
          "status": {
            "type": "string",
            "description": "Flight status description"
          },
          "aircraftType": {
            "type": "string"
          },
          "routeDistance": {
            "type": "number",
            "description": "Planned flight distance in statute miles"
          },
          "filedAirspeed": {
            "type": "number",
            "description": "Filed IFR airspeed in knots"
          },
          "filedAltitude": {
            "type": "number",
            "description": "Filed IFR altitude in 100s of feet"
          },
          "route": {
            "type": "string"
          },
          "baggageClaim": {
            "type": "string"
          },
          "seatsCabinBusiness": {
            "type": "number",
            "description": "Business class seats"
          },
          "seatsCabinCoach": {
            "type": "number",
            "description": "Coach class seats"
          },
          "seatsCabinFirst": {
            "type": "number",
            "description": "First class seats"
          },
          "gateOrigin": {
            "type": "string"
          },
          "gateDestination": {
            "type": "string"
          },
          "terminalOrigin": {
            "type": "string"
          },
          "terminalDestination": {
            "type": "string"
          },
          "scheduledOut": {
            "format": "date-time",
            "type": "string"
          },
          "estimatedOut": {
            "format": "date-time",
            "type": "string"
          },
          "actualOut": {
            "format": "date-time",
            "type": "string"
          },
          "scheduledOff": {
            "format": "date-time",
            "type": "string"
          },
          "estimatedOff": {
            "format": "date-time",
            "type": "string"
          },
          "actualOff": {
            "format": "date-time",
            "type": "string"
          },
          "scheduledOn": {
            "format": "date-time",
            "type": "string"
          },
          "estimatedOn": {
            "format": "date-time",
            "type": "string"
          },
          "actualOn": {
            "format": "date-time",
            "type": "string"
          },
          "scheduledIn": {
            "format": "date-time",
            "type": "string"
          },
          "estimatedIn": {
            "format": "date-time",
            "type": "string"
          },
          "actualIn": {
            "format": "date-time",
            "type": "string"
          },
          "foresightPredictionsAvailable": {
            "type": "boolean",
            "description": "Foresight predictions available"
          },
          "blocked": {
            "type": "boolean"
          },
          "diverted": {
            "type": "boolean"
          },
          "cancelled": {
            "type": "boolean"
          },
          "positionOnly": {
            "type": "boolean"
          },
          "originCode": {
            "type": "string"
          },
          "originCodeIcao": {
            "type": "string"
          },
          "originCodeIata": {
            "type": "string"
          },
          "originCodeLid": {
            "type": "string"
          },
          "originTimezone": {
            "type": "string"
          },
          "originName": {
            "type": "string"
          },
          "originCity": {
            "type": "string"
          },
          "destinationCode": {
            "type": "string"
          },
          "destinationCodeIcao": {
            "type": "string"
          },
          "destinationCodeIata": {
            "type": "string"
          },
          "destinationCodeLid": {
            "type": "string"
          },
          "destinationTimezone": {
            "type": "string"
          },
          "destinationName": {
            "type": "string"
          },
          "destinationCity": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "ident",
          "status",
          "foresightPredictionsAvailable",
          "blocked",
          "diverted",
          "cancelled",
          "positionOnly",
          "createdAt",
          "updatedAt"
        ]
      },
      "FlightResponseEntity": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "flightNo": {
            "type": "string"
          },
          "callSign": {
            "type": "string"
          },
          "aircraft": {
            "$ref": "#/components/schemas/AircraftEntity"
          },
          "airline": {
            "$ref": "#/components/schemas/PartialAirlineEntity"
          },
          "arrival": {
            "$ref": "#/components/schemas/RouteInfoEntity"
          },
          "departure": {
            "$ref": "#/components/schemas/RouteInfoEntity"
          },
          "cargo": {
            "type": "boolean"
          },
          "greatCircleDistance": {
            "$ref": "#/components/schemas/StringifiedGreatCircleDistanceEntity"
          },
          "date": {
            "format": "date-time",
            "type": "string"
          },
          "image": {
            "type": "string"
          },
          "flightAwareData": {
            "$ref": "#/components/schemas/FlightAwareDataEntity"
          }
        },
        "required": [
          "id",
          "flightNo",
          "callSign",
          "airline",
          "arrival",
          "departure",
          "cargo",
          "greatCircleDistance",
          "date"
        ]
      },
...
      "FlightPositionEntity": {
        "type": "object",
        "properties": {
          "fa_flight_id": {
            "type": "string",
            "description": "FlightAware flight ID for this position"
          },
          "altitude": {
            "type": "number",
            "description": "Aircraft altitude in hundreds of feet",
            "example": 350
          },
          "altitude_change": {
            "type": "string",
            "description": "Altitude change indicator",
            "enum": [
              "C",
              "D",
              "-"
            ],
            "example": "C"
          },
          "groundspeed": {
            "type": "number",
            "description": "Groundspeed in knots",
            "example": 450
          },
          "heading": {
            "type": "number",
            "description": "Aircraft heading in degrees (0-360)",
            "example": 270,
            "minimum": 0,
            "maximum": 360
          },
          "latitude": {
            "type": "number",
            "description": "Latitude position",
            "example": 40.7128
          },
          "longitude": {
            "type": "number",
            "description": "Longitude position",
            "example": -74.006
          },
          "timestamp": {
            "type": "string",
            "description": "Time that position was received",
            "example": "2024-03-20T14:30:00Z"
          },
          "update_type": {
            "type": "string",
            "description": "Update type: P=projected, O=oceanic, Z=radar, A=ADS-B, M=multilateration, D=datalink, X=surface and near surface, S=space-based, V=virtual event",
            "enum": [
              "P",
              "O",
              "Z",
              "A",
              "M",
              "D",
              "X",
              "S",
              "V"
            ]
          }
        },
        "required": [
          "altitude",
          "altitude_change",
          "groundspeed",
          "latitude",
          "longitude",
          "timestamp"
        ]
      },
      "FlightTrackResponseEntity": {
        "type": "object",
        "properties": {
          "actual_distance": {
            "type": "number",
            "description": "Distance flown in miles as of the latest position point",
            "example": 2450
          },
          "positions": {
            "description": "Array of flight positions",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FlightPositionEntity"
            }
          }
        },
        "required": [
          "positions"
        ]
      },
...
    }
  }
}